create pages, loja, sobre
