+++ src/App.css
@@ -29,91 +29,3 @@ body, input, button{
   font-family: Industrial, Arial, Helvetica, sans-serif;
   font-size: 14px;
 }
-/* 
-.container {
-  width: 100%;
-  height: 100%;
-  display: flexbox; 
-} */
-.container2 {
-  margin: 150px 200px auto;
-  max-width: 650px;
-  width: 100%;
-  display: flex; 
-}
-
-@media screen and (max-width: 1024px) {  
-  .container2 {
-    margin: 150px 0 auto;
-    max-width: 650px;
-    width: 100%;
-    display: flex; 
-  }
-}
-
-@media screen and (max-width: 520px) {
-  
-  .direita {      
-    width: 40%;
-    height: 60%; 
-    bottom: 0;
-    right: 0;                 
-  }
-  
-  .direita img {
-    max-width: 100%;  
-  }
-  .container2 {
-    margin: 150px 0 auto;
-    max-width: 650px;
-    width: 100%;
-    display: flex; 
-  }
-}
-
-@media screen and (max-width: 450px) {
-  .content {
-    width: 100%; 
-    padding: 20px;    
-  }
-  
-  .content > p {
-    font-size: 30px;
-    line-height: 22px;
-    margin-bottom: 20px;
-    font-style: normal;
-    color: #FFF;
-    text-align: center;
-  }
-  
-  .content h1 {
-    font-size: 70px;
-    color: #f54900;
-    letter-spacing: 0.7rem;
-    font-weight: 900;
-    margin-bottom: 20px;
-    text-align: center;
-  }
-  
-  .direita { display: none; }  
-  .container2 {
-    margin: 150px 0 auto;
-    max-width: 450px;
-    width: 100%;
-    display: flex; 
-  }
-
-  .content button.btn {
-    border: solid 3px #f54900;
-    border-radius: 10px;
-    max-width: 450px;
-    width: 100%;
-    height: 50px;
-    padding: 0 20px;
-    font-size: 25px;
-    font-weight: bold;
-    color: #FFF;
-    background: none;
-    cursor: pointer;  
-  }
-}
