+++ src/App.js
@@ -8,10 +8,8 @@ import './App.css';
 function App() {
   return (
     <>
-      <Header />     
-      <div className="container2">
-        <Routes />
-      </div>    
+      <Header />       
+      <Routes />
     </>
   );
 }
