+++ src/pages/Home/Home.css
@@ -1,3 +1,9 @@
+.container {
+  margin: 150px 0 auto 200px;
+  max-width: 650px;
+  display: flex; 
+}
+
 .direita {
   position: absolute;  
   width: 40%;
@@ -5,7 +11,7 @@
   bottom: 0;
   right: 0;                 
 }
-  
+
 .direita img {
   max-width: 100%;
   height: 100%;
@@ -16,7 +22,7 @@
   width: 100%; 
   padding: 20px;
 }
-  
+
 .content > p {
   font-size: 30px;
   line-height: 22px;
@@ -54,8 +60,23 @@
   transition: 0.1s;
 }
 
+@media screen and (max-width: 1024px) {  
+  .container {
+    margin: 150px 0 auto;
+    max-width: 650px;
+    width: 100%;
+    display: flex; 
+  }
+}
 
-@media screen and (max-width: 520px) {  
+@media screen and (max-width: 520px) {
+  .container {
+    margin: 150px 0 auto;
+    max-width: 650px;
+    width: 100%;
+    display: flex; 
+  }
+
   .direita {      
     width: 40%;
     height: 60%; 
@@ -68,7 +89,7 @@
   }
 }
 
-@media screen and (max-width: 450px) {
+@media screen and (max-width: 500px) {
   .content {
     width: 100%; 
     padding: 20px;
